refactor(i18n): migrate to react-i18next and clean up translation logic
(close #245)
