--- apps/movies-ui/src/i18n/TranslationKeys.ts
@@ -1,1 +1,0 @@
-export type TranslationKeys = Record<string, Record<string, string>>;
